log in successfully, database need to be updated (or wiped)

+++ src/main/java/com/ajax/model/Person.java
@@ -52,7 +52,7 @@ public class Person {
         return phone;
     }
 
-    public void setPhone(int phone) {
+    public void setPhone(long phone) {
         this.phone = phone;
     }
